@@ -9,28 +9,28 @@
     "eject": "react-scripts eject"
   },
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.3.1",
-    "@mui/lab": "^5.0.0-alpha.66",
-    "@mui/material": "^5.3.1",
-    "@mui/styles": "^5.3.0",
-    "@mui/system": "^5.3.0",
-    "@mui/utils": "^5.3.0",
-    "@mui/x-data-grid": "^5.4.0",
-    "@reduxjs/toolkit": "^1.7.1",
-    "nanoid": "^3.2.0",
-    "notistack": "^2.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
-    "recoil": "^0.6.1",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/lab": "^5.0.0-alpha.86",
+    "@mui/material": "^5.8.4",
+    "@mui/styles": "^5.8.4",
+    "@mui/system": "^5.8.4",
+    "@mui/utils": "^5.8.4",
+    "@mui/x-data-grid": "^5.12.2",
+    "@reduxjs/toolkit": "^1.8.2",
+    "nanoid": "^4.0.0",
+    "notistack": "^2.0.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "recoil": "^0.7.3",
     "web-vitals": "^2.1.4"
   },
   "devDependencies": {
-    "prettier": "^2.5.1"
+    "prettier": "^2.7.1"
   },
   "eslintConfig": {
     "extends": [
